requirements.txt
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-cleanup==6.0.0
django-templated-mail==1.1.1
django-cloudinary-storage==0.3.0
django-celery-email==3.0.0
django-celery-beat==2.4.0
django-pickling==1.0
django-redis==5.2.0
django-cacheops==6.1
django-money==3.0.0
django-svg-image-form-field==1.0.1
django-uuslug==2.0.0

djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-yasg==1.21.4
drf_base64==2.0
djangochannelsrestframework==1.1.0

channels==4.0.0
channels-redis==4.0.0
asgiref==3.5.2
orjson==3.8.3
psycopg2-binary==2.9.5
redis==4.4.0
Pillow==9.3.0
whitenoise==6.2.0
celery==5.2.7
gevent==22.10.2
argon2-cffi==21.3.0
pydantic[email]==1.10.2
sentry-sdk==1.11.1
py-moneyed==2.0
django-silk==5.0.2
daphne==4.0.0
Twisted[http2]==22.10.0
gunicorn==20.1.0

parameterized==0.8.1
factory_boy==3.2.1
Faker==15.3.4
coverage==6.5.0
